Move AutoStructured to a separate file (#2936)
* Move AutoStructured to a separate file

* lint
